% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abun.R
\name{plot_abun}
\alias{plot_abun}
\title{Plot functional group abundance}
\usage{
plot_abun(fg, grp_list, atlantis_outputs, ini.pol, end.pol)
}
\arguments{
\item{end.pol}{}
}
\value{
plot_out
}
\description{
Plot functional group abundance
}
