% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Diets.R
\name{plot_Diets}
\alias{plot_Diets}
\title{Plot diets}
\usage{
plot_Diets(
  dietsAll,
  FG_to_plot,
  threshold,
  yearsselected,
  grp_list,
  atlantis_outputs,
  startyear
)
}
\arguments{
\item{startyear}{}
}
\value{
dietplot
}
\description{
Plot diets
}
