% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Preds.R
\name{plot_Preds}
\alias{plot_Preds}
\title{Plot diet of predators}
\usage{
plot_Preds(
  dietsAll,
  FG_to_plot,
  threshold,
  yearsselected,
  grp_list,
  timeperiod,
  startyear
)
}
\arguments{
\item{startyear}{}
}
\value{
dietplot
}
\description{
Plot diet of predators
}
