% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rn_vs_sn.R
\name{plot_rn_vs_sn}
\alias{plot_rn_vs_sn}
\title{Plot reserve and structural Nitrogen}
\usage{
plot_rn_vs_sn(fg, grp_list, atlantis_outputs)
}
\arguments{
\item{atlantis_outputs}{}
}
\value{
plot_out
}
\description{
Plot reserve and structural Nitrogen
}
