% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Plot spatial distribution and center-of-gravity of a functional group}
\usage{
plot_spatial(fg, spatial_plot_type, grp_list, atlantis_outputs, emocc_sf)
}
\arguments{
\item{atlantis_outputs}{}
}
\value{
out_plot
}
\description{
Plot spatial distribution and center-of-gravity of a functional group
}
