% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wage_timeseries.R
\name{plot_wage_timeseries}
\alias{plot_wage_timeseries}
\title{Plot weight-at-age time series}
\usage{
plot_wage_timeseries(fg, grp_list, atlantis_outputs)
}
\arguments{
\item{atlantis_outputs}{}
}
\value{
plot_out
}
\description{
Plot weight-at-age time series
}
