% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wage_timeseries_old.R
\name{plot_wage_timeseries_old}
\alias{plot_wage_timeseries_old}
\title{deprecated (old) weight at age calculation}
\usage{
plot_wage_timeseries_old(fg, atlantis_outputs, grp_list)
}
\arguments{
\item{grp_list}{}
}
\value{
plot_out
}
\description{
just keeping it here in case we need to revert for some reason
}
